+++ core/analysis/color/palette_analysis.py
@@ -77,7 +77,7 @@ class ColorPaletteAnalysis(IAnalysisJob):
         model = None
 
         if self.coverage is not None:
-            self.resolution = (stop - start) / ((stop - start) * self.coverage)
+            self.resolution = int((stop - start) / ((stop - start) * self.coverage))
 
         for i in range(start, stop + 1, self.resolution):
             sign_progress((c - start) / ((stop - start) + 1))
